adjust tests for new scoreboard helpers
