+++ lib/game.js
@@ -122,7 +122,6 @@ Game.prototype.showGameOverScreen = function(){
 
 Game.prototype.sendHighScore = function(){
     scoreboard.sendHighScore(points);
-    // points = 0;
 };
 
 Game.prototype.resetPoints = function(){
