+++ lib/scoreboard.js
@@ -17,21 +17,17 @@ Scoreboard.prototype.showHighScoreEntry = function(points){
 };
 
 Scoreboard.prototype.sendHighScore = function(points){
-    scoreboardRecords = helpers.addToScoreboardRecords(scoreboardRecords, points);
-    fireDb.ref('highscore/').set({
-      highscores: scoreboardRecords
-    });
-    event.preventDefault();
-    $(".high-score-entry").hide();
-    loadScoreboard();
-    $('.scoreboard').show();
+  scoreboardRecords = helpers.addToScoreboardRecords(scoreboardRecords, points);
+  fireDb.ref('highscore/').set({
+    highscores: scoreboardRecords
+  });
+  event.preventDefault();
+  $(".high-score-entry").hide();
+  this.loadScoreboard();
+  $('.scoreboard').show();
 };
 
 Scoreboard.prototype.loadScoreboard = function(){
-  loadScoreboard();
-};
-
-function loadScoreboard(){
   scoreboardRecords = [];
   $('#scoreboard-records').empty();
 
@@ -40,52 +36,6 @@ function loadScoreboard(){
     scoreboardRecords = helpers.addScores(scoreboardRecords, allScores);
     helpers.renderScores(allScores, scoreboardRecords.length);
   });
-}
-
-// function sortScores(scores){
-//   var sortedScores = scores.sort(function(a, b){return b.points - a.points;});
-//   return sortedScores;
-// }
-
-// Scoreboard.prototype.addToScoreboardRecords = function (points) {
-//   var username = $("#username").val() || "unknown user";
-//   scoreboardRecords.push({username: username, points: points});
-//   return scoreboardRecords;
-// };
-//
-// Scoreboard.prototype.sortScores = function(scores){
-//   var sortedScores = scores.sort(function(a, b){return b.points - a.points;});
-//   return sortedScores;
-// };
-
-// Scoreboard.prototype.addScores = function(allScores){
-//   for (var i = 0; i < 5; i++) {
-//     if (allScores[i] !== undefined) {
-//       scoreboardRecords.push(allScores[i]);
-//     }
-//   }
-//   return scoreboardRecords;
-// };
-
-// Scoreboard.prototype.renderScores = function(allScores, scoreboardLength){
-//   // var results = ""
-//   for (var i = 0; i < scoreboardLength; i++) {
-//     $('#scoreboard-records').append(i+1 + ". " + allScores[i].username + ": " + allScores[i].points + "<br>");
-//     // results += (i+1) + ". " + allScores[i].username + ": " + allScores[i].points + "<br>"
-//   }
-//   // $("#scoreboard-records").html(results)
-// };
-
-// Scoreboard.prototype.loadScoreboard = function(){
-//   var scoreboard = this;
-//   scoreboardRecords = [];
-//   $('#scoreboard-records').empty();
-//
-//   fireDb.ref('highscore/').once('value').then(function(scores){
-//     var allScores = scoreboard.sortScores(scores.val().highscores);
-//     scoreboard.addScores(allScores);
-//     scoreboard.renderScores(allScores, scoreboardRecords.length);
-//   });
-// };
+};
 
 module.exports = Scoreboard;
