+++ lib/sushi.js
@@ -1,7 +1,7 @@
-var eggRoll = document.getElementById('egg-roll')
-var roeRoll = document.getElementById('roe-roll')
-var fishRoll = document.getElementById('fish-roll')
-var pictures = [eggRoll, roeRoll, fishRoll]
+var eggRoll = document.getElementById('egg-roll');
+var roeRoll = document.getElementById('roe-roll');
+var fishRoll = document.getElementById('fish-roll');
+var pictures = [eggRoll, roeRoll, fishRoll];
 
 function Sushi(options) {
   var rowsForSprites = [70, 170, 270, 370];
