+++ lib/trash.js
@@ -1,6 +1,6 @@
-var poop = document.getElementById('kawaii-poop')
-var toaster = document.getElementById('kawaii-toaster')
-var pictures = [poop, toaster]
+var poop = document.getElementById('kawaii-poop');
+var toaster = document.getElementById('kawaii-toaster');
+var pictures = [poop, toaster];
 
 function Trash(options) {
   var rowsForTrash = [70, 170, 270, 370];
